requirements.txt
# Core Azure SDKs
azure-identity==1.23.0
azure-mgmt-resource==23.1.0
azure-mgmt-consumption==10.0.0
azure-mgmt-subscription==3.1.1
azure-mgmt-resourcegraph==8.0.0
azure-core==1.34.0
azure-common==1.1.28
azure-mgmt-reservations==2.3.0


# Email & Environment
sendgrid==6.12.2
python-dotenv==1.1.0
requests==2.32.3

# CLI / Formatting / Reporting
tabulate==0.9.0

# Testing (optional)
pytest==8.3.5

# Required transitive dependencies (kept if you freeze with pip freeze)
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
